% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_delcampe_export.R
\name{build_template_description}
\alias{build_template_description}
\title{Build template description from title and standard text}
\usage{
build_template_description(title)
}
\arguments{
\item{title}{Character string with extracted title}
}
\value{
Character string with formatted description
}
\description{
Build template description from title and standard text
}
