% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_database.R
\name{calculate_image_hash}
\alias{calculate_image_hash}
\title{Calculate MD5 hash of an image file for deduplication}
\usage{
calculate_image_hash(image_path)
}
\arguments{
\item{image_path}{Path to the image file}
}
\value{
Character string with MD5 hash, or NULL if error
}
\description{
Calculate MD5 hash of an image file for deduplication
}
