% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_time_helpers.R
\name{calculate_next_10am_pacific}
\alias{calculate_next_10am_pacific}
\title{Calculate next 10:00 AM Pacific time from current moment}
\usage{
calculate_next_10am_pacific()
}
\value{
POSIXct object in UTC timezone representing next 10:00 AM Pacific
}
\description{
Returns the next occurrence of 10:00 AM Pacific Time (PDT/PST).
R automatically handles daylight saving time transitions when using
the "America/Los_Angeles" timezone.
}
\examples{
next_10am <- calculate_next_10am_pacific()
format(next_10am, "\%Y-\%m-\%d \%H:\%M:\%S \%Z")
}
