% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_database.R
\name{copy_existing_crops}
\alias{copy_existing_crops}
\title{Copy existing crop files to a new directory}
\usage{
copy_existing_crops(source_paths, dest_dir)
}
\arguments{
\item{source_paths}{Vector of paths to copy from}

\item{dest_dir}{Directory to copy to}
}
\value{
List with new_paths and success
}
\description{
Copy existing crop files to a new directory
}
