% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_policy_helper.R
\name{create_default_ebay_policies}
\alias{create_default_ebay_policies}
\title{Create Default eBay Business Policies}
\usage{
create_default_ebay_policies(ebay_api)
}
\arguments{
\item{ebay_api}{An authenticated eBay API object from mod_ebay_auth_server}
}
\value{
A list with created policy IDs
}
\description{
Creates default business policies for eBay sandbox/production if none exist
}
