% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgur_upload.R
\name{delete_from_imgur}
\alias{delete_from_imgur}
\title{Delete Image from Imgur}
\usage{
delete_from_imgur(delete_hash, client_id = Sys.getenv("IMGUR_CLIENT_ID"))
}
\arguments{
\item{delete_hash}{Delete hash from upload response}

\item{client_id}{Imgur Client ID}
}
\description{
Delete Image from Imgur
}
