% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_category_config.R
\docType{data}
\name{EBAY_POSTCARD_CATEGORIES}
\alias{EBAY_POSTCARD_CATEGORIES}
\title{eBay Postcard Leaf Categories}
\format{
An object of class \code{list} of length 2.
}
\usage{
EBAY_POSTCARD_CATEGORIES
}
\description{
eBay Postcard Leaf Categories
}
\keyword{datasets}
