% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_time_helpers.R
\name{format_ebay_schedule_time}
\alias{format_ebay_schedule_time}
\title{Format datetime for eBay ScheduleTime field (ISO 8601)}
\usage{
format_ebay_schedule_time(datetime)
}
\arguments{
\item{datetime}{POSIXct object (will be converted to UTC if not already)}
}
\value{
Character string in format YYYY-MM-DDTHH:MM:SS.SSSZ
}
\description{
Converts a POSIXct datetime to eBay's required ISO 8601 format with
milliseconds and Z suffix (UTC indicator).
}
\examples{
schedule_time <- Sys.time() + 7200  # 2 hours from now
format_ebay_schedule_time(schedule_time)
# Returns: "2025-11-01T19:00:00.000Z"
}
