% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_sync_helpers.R
\name{get_ebay_user_id_from_session}
\alias{get_ebay_user_id_from_session}
\title{Get eBay user ID from session}
\usage{
get_ebay_user_id_from_session(con, session_id)
}
\arguments{
\item{con}{Database connection}

\item{session_id}{Session ID}
}
\value{
Character eBay user ID or NULL
}
\description{
Get eBay user ID from session
}
