% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_database.R
\name{get_posting_statistics}
\alias{get_posting_statistics}
\title{Get statistics on eBay postings}
\usage{
get_posting_statistics(session_id = NULL)
}
\arguments{
\item{session_id}{Optional session filter}
}
\value{
List with statistics
}
\description{
Get statistics on eBay postings
}
