% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_time_helpers.R
\name{get_romania_from_utc}
\alias{get_romania_from_utc}
\title{Convert UTC to Romania time}
\usage{
get_romania_from_utc(utc_time)
}
\arguments{
\item{utc_time}{POSIXct object in UTC timezone}
}
\value{
POSIXct object in Europe/Bucharest timezone
}
\description{
Helper function to convert UTC time to Europe/Bucharest timezone.
R automatically handles daylight saving time transitions.
}
\examples{
utc_time <- as.POSIXct("2025-11-15 17:00:00", tz = "UTC")
romania_time <- get_romania_from_utc(utc_time)
format(romania_time, "\%Y-\%m-\%d \%H:\%M \%Z")
# Returns: "2025-11-15 19:00 EET"
}
