% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_category_config.R
\name{get_stamp_category}
\alias{get_stamp_category}
\title{Get appropriate stamp category based on data}
\usage{
get_stamp_category(ai_data)
}
\arguments{
\item{ai_data}{List with AI extraction data (year, country, etc.)}
}
\value{
Numeric category ID (leaf category)
}
\description{
Get appropriate stamp category based on data
}
