% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_database.R
\name{get_tracking_data}
\alias{get_tracking_data}
\title{Get tracking data with filters}
\usage{
get_tracking_data(date_filter = "", ebay_filter = "")
}
\arguments{
\item{date_filter}{SQL WHERE clause for date filtering}

\item{ebay_filter}{SQL WHERE clause for eBay status filtering}
}
\value{
Data frame with tracking data
}
\description{
Query processing history with date and eBay status filters
}
