% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_database.R
\name{get_tracking_statistics}
\alias{get_tracking_statistics}
\title{Get tracking statistics - FIXED VERSION}
\usage{
get_tracking_statistics()
}
\value{
List with tracking statistics
}
\description{
Get tracking statistics - FIXED VERSION
}
