% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_database_extension.R
\name{initialize_ebay_tables}
\alias{initialize_ebay_tables}
\title{Initialize eBay listings table}
\usage{
initialize_ebay_tables(db_path = NULL)
}
\description{
Initialize eBay listings table
}
