% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_system.R
\name{list_users}
\alias{list_users}
\title{List Users}
\usage{
list_users(current_user_role)
}
\arguments{
\item{current_user_role}{Character string - role of user requesting list}
}
\value{
Data frame with user information (excluding password hashes)
}
\description{
Retrieves list of users with role-based filtering.
Admins/masters see all users. Regular users see only themselves.
}
\examples{
\dontrun{
# Admin view - sees all users
users <- list_users("admin")

# User view - sees only self
users <- list_users("user")
}

}
