% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_sync_helpers.R
\name{log_sync_complete}
\alias{log_sync_complete}
\title{Log completion of sync operation}
\usage{
log_sync_complete(con, sync_id, items_synced, api_calls)
}
\arguments{
\item{con}{Database connection}

\item{sync_id}{Sync log ID}

\item{items_synced}{Number of items synced}

\item{api_calls}{Number of API calls made}
}
\description{
Log completion of sync operation
}
