% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_sync_helpers.R
\name{log_sync_error}
\alias{log_sync_error}
\title{Log sync error}
\usage{
log_sync_error(con, sync_id, error_message)
}
\arguments{
\item{con}{Database connection}

\item{sync_id}{Sync log ID}

\item{error_message}{Error message}
}
\description{
Log sync error
}
