% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_sync_helpers.R
\name{log_sync_start}
\alias{log_sync_start}
\title{Log start of sync operation}
\usage{
log_sync_start(con, ebay_user_id)
}
\arguments{
\item{con}{Database connection}

\item{ebay_user_id}{eBay user ID}
}
\value{
Integer sync_id
}
\description{
Log start of sync operation
}
