% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_helpers.R
\name{map_condition_to_trading_id}
\alias{map_condition_to_trading_id}
\title{Map Condition to Trading API ConditionID}
\usage{
map_condition_to_trading_id(condition)
}
\arguments{
\item{condition}{Condition string from AI}
}
\value{
Integer ConditionID for postcards/collectibles
}
\description{
Trading API uses integer ConditionID instead of text codes
Reference: https://developer.ebay.com/devzone/finding/callref/enums/conditionIdList.html
}
