% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_stamp_helpers.R
\name{map_stamp_grade_to_ebay}
\alias{map_stamp_grade_to_ebay}
\title{Map Stamp Grade to eBay Condition}
\usage{
map_stamp_grade_to_ebay(grade)
}
\arguments{
\item{grade}{Character: Grade from AI extraction (MNH, MH, Used, Unused)}
}
\value{
Character: eBay condition code
}
\description{
Converts philatelic grade to eBay condition code
}
