% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_database.R
\name{mark_processing_reused}
\alias{mark_processing_reused}
\title{Mark processing as reused in database}
\usage{
mark_processing_reused(
  current_session_id,
  source_session_id,
  image_id,
  source_image_id
)
}
\arguments{
\item{current_session_id}{Current session ID}

\item{source_session_id}{Source session ID}

\item{image_id}{Current image ID}

\item{source_image_id}{Source image ID}
}
\value{
TRUE if successful
}
\description{
Mark processing as reused in database
}
