% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_delcampe_export.R
\name{mod_delcampe_export_server}
\alias{mod_delcampe_export_server}
\title{Delcampe Export Server Functions}
\usage{
mod_delcampe_export_server(
  id,
  image_paths = reactive(NULL),
  image_type = "combined"
)
}
\arguments{
\item{image_paths}{Reactive containing vector of image paths}

\item{image_type}{Character string ("lot" or "combined") to distinguish image types}
}
\description{
Delcampe Export Server Functions
}
