% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_delcampe_export.R
\name{mod_delcampe_export_ui}
\alias{mod_delcampe_export_ui}
\title{Delcampe Export UI Function}
\usage{
mod_delcampe_export_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module for handling Delcampe export actions.
Provides UI for sending processed images to Delcampe with AI extraction support.
}
