% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ebay_auth.R
\name{mod_ebay_auth_server}
\alias{mod_ebay_auth_server}
\title{eBay Authentication Server Module}
\usage{
mod_ebay_auth_server(id, parent_session = NULL)
}
\description{
eBay Authentication Server Module
}
