% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ebay_listings.R
\name{mod_ebay_listings_server}
\alias{mod_ebay_listings_server}
\title{eBay Listings Viewer Server}
\usage{
mod_ebay_listings_server(id, ebay_api, session_id, ebay_account_manager = NULL)
}
\arguments{
\item{id}{Module ID}

\item{ebay_api}{Reactive returning EbayAPI object (with $trading field)}

\item{session_id}{Reactive returning session ID}
}
\description{
eBay Listings Viewer Server
}
