% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ebay_postcard.R
\name{mod_ebay_postcard_server}
\alias{mod_ebay_postcard_server}
\title{eBay Postcard Listing Server Module}
\usage{
mod_ebay_postcard_server(id, ebay_api)
}
\description{
eBay Postcard Listing Server Module
}
