% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_postal_card_processor.R
\name{mod_postal_card_processor_server}
\alias{mod_postal_card_processor_server}
\title{postal_card_processor Server Functions}
\usage{
mod_postal_card_processor_server(
  id,
  card_type = "face",
  on_grid_update = NULL,
  on_image_upload = NULL,
  on_extraction_complete = NULL
)
}
\description{
postal_card_processor Server Functions
}
