% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_postal_card_processor.R
\name{mod_postal_card_processor_ui}
\alias{mod_postal_card_processor_ui}
\title{postal_card_processor UI Function}
\usage{
mod_postal_card_processor_ui(id, card_type = "face")
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module for processing postal cards with draggable gridlines and Python integration.
}
