% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings_server.R
\name{mod_settings_server}
\alias{mod_settings_server}
\title{Settings Module Server - FIXED VERSION}
\usage{
mod_settings_server(id, current_user)
}
\description{
Settings Module Server - FIXED VERSION
}
