% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stamp_face_processor.R
\name{mod_stamp_face_processor_server}
\alias{mod_stamp_face_processor_server}
\title{stamp_face_processor Server Functions}
\usage{
mod_stamp_face_processor_server(
  id,
  stamp_type = "face",
  on_grid_update = NULL,
  on_image_upload = NULL,
  on_extraction_complete = NULL
)
}
\description{
stamp_face_processor Server Functions
}
