% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stamp_face_processor.R
\name{mod_stamp_face_processor_ui}
\alias{mod_stamp_face_processor_ui}
\title{stamp_face_processor UI Function}
\usage{
mod_stamp_face_processor_ui(id, stamp_type = "face")
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module for processing stamp face images with draggable gridlines and Python integration.
}
