% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stamp_verso_processor.R
\name{mod_stamp_verso_processor_server}
\alias{mod_stamp_verso_processor_server}
\title{stamp_verso_processor Server Functions}
\usage{
mod_stamp_verso_processor_server(
  id,
  stamp_type = "verso",
  on_grid_update = NULL,
  on_image_upload = NULL,
  on_extraction_complete = NULL
)
}
\description{
stamp_verso_processor Server Functions
}
