% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tracking_viewer.R
\name{mod_tracking_viewer_ui}
\alias{mod_tracking_viewer_ui}
\title{tracking_viewer UI Function}
\usage{
mod_tracking_viewer_ui(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\description{
A shiny Module for viewing processing tracking data.
}
