% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_sync_helpers.R
\name{parse_seller_list_response}
\alias{parse_seller_list_response}
\title{Parse GetSellerList XML response}
\usage{
parse_seller_list_response(xml_response)
}
\arguments{
\item{xml_response}{Character XML response from eBay}
}
\value{
List with items and has_more_items
}
\description{
Parse GetSellerList XML response
}
\keyword{internal}
