% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_stamp_categories.R
\docType{data}
\name{STAMP_CATEGORIES}
\alias{STAMP_CATEGORIES}
\title{Complete eBay Stamp Category Hierarchy}
\format{
An object of class \code{list} of length 16.
}
\usage{
STAMP_CATEGORIES
}
\description{
Nested list structure matching eBay's category hierarchy.
Each region contains countries with their category IDs.
}
\keyword{datasets}
