% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_database.R
\name{start_processing_session}
\alias{start_processing_session}
\title{Start processing session - FIXED VERSION}
\usage{
start_processing_session(session_id, user_id, session_type = "general")
}
\arguments{
\item{session_id}{Session identifier}

\item{user_id}{User identifier}

\item{session_type}{Type of session (ignored)}
}
\value{
Session ID
}
\description{
Start processing session - FIXED VERSION
}
