% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_database.R
\name{track_processing_action}
\alias{track_processing_action}
\title{Track processing action - FIXED VERSION}
\usage{
track_processing_action(
  image_id,
  action,
  user_id,
  details = NULL,
  success = TRUE,
  error_message = NULL
)
}
\arguments{
\item{image_id}{Image ID}

\item{action}{Action performed}

\item{user_id}{User performing action}

\item{details}{Additional details}

\item{success}{Whether action succeeded}

\item{error_message}{Error message if failed}
}
\value{
Success status
}
\description{
Track processing action - FIXED VERSION
}
