% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_database_extension.R
\name{update_ebay_listing_status}
\alias{update_ebay_listing_status}
\title{Update eBay listing status}
\usage{
update_ebay_listing_status(sku, status, error_message = NULL)
}
\description{
Update eBay listing status
}
