% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebay_sync_helpers.R
\name{update_listings_cache}
\alias{update_listings_cache}
\title{Update cached eBay data in database (INSERT or UPDATE)}
\usage{
update_listings_cache(con, ebay_items)
}
\arguments{
\item{con}{Database connection}

\item{ebay_items}{List of items from parse_seller_list_response}
}
\description{
Update cached eBay data in database (INSERT or UPDATE)
}
