% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgur_upload.R
\name{upload_to_imgbb}
\alias{upload_to_imgbb}
\title{Upload Image to imgbb}
\usage{
upload_to_imgbb(image_path, api_key = Sys.getenv("IMGBB_API_KEY"))
}
\arguments{
\item{image_path}{Local file path to image}

\item{api_key}{imgbb API key (get from https://api.imgbb.com/)}
}
\value{
List with success, url, delete_url
}
\description{
Upload an image file to imgbb and get a public HTTPS URL
}
