% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgur_upload.R
\name{upload_to_imgur}
\alias{upload_to_imgur}
\title{Upload Image to Imgur (DEPRECATED - use imgbb instead)}
\usage{
upload_to_imgur(image_path, client_id = Sys.getenv("IMGUR_CLIENT_ID"))
}
\arguments{
\item{image_path}{Local file path to image}

\item{client_id}{Imgur Client ID (get from https://api.imgur.com/oauth2/addclient)}
}
\value{
List with success, url, delete_hash
}
\description{
Upload an image file to Imgur and get a public HTTPS URL
}
