% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_database.R
\name{validate_existing_crops}
\alias{validate_existing_crops}
\title{Validate that crop files from previous processing still exist}
\usage{
validate_existing_crops(cropped_paths)
}
\arguments{
\item{cropped_paths}{Vector of file paths to validate}
}
\value{
List with all_exist (logical) and missing_files (vector)
}
\description{
Validate that crop files from previous processing still exist
}
