% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tracking_viewer.R
\name{mod_tracking_viewer_server}
\alias{mod_tracking_viewer_server}
\title{Enhanced Tracking Viewer Server}
\usage{
mod_tracking_viewer_server(id)
}
\arguments{
\item{id}{Module ID}
}
\description{
Server logic for tracking viewer with DT::datatable
}
